Support encoding of ELF file header
A lot of logic for a small change in any other language (really!).
This change adds supports for encoding an ELF file header. In
reality, most of the 'work' here is setting up a basic scheme
structure system for creation and encoding of dictionaries. This
attempts to walk a fine line for adding requirements to minischeme
versus really slow logic due to poor infrastructure.